Extract GeoJson to separate module
