+++ js/src/Marker.js
@@ -185,53 +185,6 @@ export const MarkerLayerView = GMapsLayerView.extend({
 })
 
 
-export const GeoJsonFeatureView = GMapsLayerView.extend({
-
-    // nameInView -> name_in_model
-    styleProperties: [
-          ['fillColor', 'fill_color'],
-          ['fillOpacity', 'fill_opacity'],
-          ['strokeColor', 'stroke_color'],
-          ['strokeOpacity', 'stroke_opacity'],
-          ['strokeWeight', 'stroke_weight']
-    ],
-
-    render() {
-        this.modelEvents() ;
-        this.geojson = this.model.get("feature")
-        const style = this.styleProperties.reduce(
-            (acc, [nameInView, nameInModel]) => {
-              return {...acc, [nameInView]: this.model.get(nameInModel)}
-            },
-            {}
-        )
-        this.geojson.properties =
-            this.geojson.properties ? this.geojson.properties : {}
-        this.geojson.properties.style = style
-    },
-
-    addToMapView(mapView) {
-        this.mapView = mapView
-        mapView.map.data.addGeoJson(this.geojson)
-    },
-
-    modelEvents() {
-        this.styleProperties.forEach(([nameInView, nameInModel]) => {
-            const callback = (() => {
-                this.geojson.properties.style = {
-                    ...this.geojson.properties.style,
-                    [nameInView]: this.model.get(nameInModel)
-                }
-                this.mapView.map.data.setStyle(
-                    (feature) => feature.getProperty('style'))
-            })
-            this.model.on(`change:${nameInModel}`, callback, this)
-        })
-    }
-
-})
-
-
 export const SymbolModel = GMapsLayerModel.extend({
     defaults: _.extend({}, GMapsLayerModel.prototype.defaults, {
         _view_name: "SymbolView",
